BatchOrchestrator R6 class

Manages a group of batch_reactive_val() objects, orchestrating validation and ensuring that their values are updated together in a single atomic transaction.

Methods

Public methods

BatchOrchestrator$new()

BatchOrchestrator$add_val()

BatchOrchestrator$is_in_transaction()

BatchOrchestrator$get_timeout()

BatchOrchestrator$wait_for_transaction()

BatchOrchestrator$request_update()

BatchOrchestrator$clone()

Method new()

Initialize the transaction manager.

Usage

BatchOrchestrator$new(validation_flow = list(), timeout = 5L)

Arguments

validation_flow

(named list of functions) Each function takes two arguments: my_value (the current or proposed value for the reactive it is validating) and all_values (a named list of all other values in the transaction), and returns a valid value. The order of this list defines the order of the validation cascade.

timeout

(length-1 integer) The maximum number of seconds to wait for a transaction to complete.

Method add_val()

Add a reactive value to be managed by this coordinator.

Usage

BatchOrchestrator$add_val(val, validation_fun = NULL)

Arguments

val

(BatchReactiveVal) The BatchReactiveVal object to add.

validation_fun

(function or NULL) The validation function for this value.

Method is_in_transaction()

Check if a transaction is currently active.

Usage

BatchOrchestrator$is_in_transaction()

Returns

(length-1 logical) TRUE if a transaction is in progress.

Method get_timeout()

Get this orchestrator's timeout setting.

Usage

BatchOrchestrator$get_timeout()

Returns

(length-1 integer) The maximum number of seconds to wait for a transaction to complete.

Method wait_for_transaction()

Block until the current transaction is complete.

Usage

BatchOrchestrator$wait_for_transaction(
  timeout = self$get_timeout(),
  call = rlang::caller_env()
)

Arguments

timeout

(length-1 integer) The maximum number of seconds to wait for a transaction to complete.

call

(environment) The execution environment to mention as the source of error messages.

Method request_update()

Attempt to update managed values in a transaction.

Usage

BatchOrchestrator$request_update(
  trigger_name,
  proposed_value,
  call = rlang::caller_env()
)

Arguments

trigger_name

(length-1 character) The name of the reactive value that initiated the change.

proposed_value

(various) The new value proposed for the trigger.

call

(environment) The execution environment to mention as the source of error messages.

Method clone()

The objects of this class are cloneable with this method.

Usage

BatchOrchestrator$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.