BatchReactiveVal R6 class

An R6 class that encapsulates a standard reactiveVal. It serves as the backend for batch_reactive_val() and is managed by a BatchOrchestrator object.

Methods

Public methods

BatchReactiveVal$new()

BatchReactiveVal$get()

BatchReactiveVal$get_rv_value()

BatchReactiveVal$set()

BatchReactiveVal$set_rv_value()

BatchReactiveVal$get_name()

BatchReactiveVal$get_orchestrator()

BatchReactiveVal$clone()

Method new()

Initialize the object.

Usage

BatchReactiveVal$new(
  name,
  value = NULL,
  orchestrator = batch_orchestrator(),
  validation_fun = NULL,
  call = rlang::caller_env()
)

Arguments

name

(length-1 character) The unique name for this reactive value.

value

(various) The initial value.

orchestrator

(BatchOrchestrator or batch_reactive_val) A BatchOrchestrator object, or a batch_reactive_val that should be included in the same batch. Defaults to creating a new batch_orchestrator().

validation_fun

(function or NULL) The validation function for this value.

call

(environment) The execution environment to use for error messages.

Method get()

Get the current value, blocking if a transaction is active.

Usage

BatchReactiveVal$get(call = rlang::caller_env())

Arguments

call

(environment) The execution environment to use for error messages.

Returns

The current, consistent value stored in the internal reactiveVal.

Method get_rv_value()

Get the raw value of the internal reactiveVal.

Usage

BatchReactiveVal$get_rv_value()

Details

This is an unsafe getter intended for internal use by the orchestrator only. It does not wait for transactions to complete.

Returns

The current value stored in the internal reactiveVal.

Method set()

Request a new value for this reactive.

Usage

BatchReactiveVal$set(value, call = rlang::caller_env())

Arguments

value

(various) The new value to set.

call

(environment) The execution environment to use for error messages.

Details

This method initiates a transaction with the orchestrator.

Method set_rv_value()

Set the raw value of the internal reactiveVal.

Usage

BatchReactiveVal$set_rv_value(value)

Arguments

value

(various) The new value to set.

Details

This is an unsafe setter intended for internal use by the orchestrator only. It does not trigger a validation cascade.

Method get_name()

Get the name of the reactive value.

Usage

BatchReactiveVal$get_name()

Returns

(length-1 character) The name of the reactive value.

Method get_orchestrator()

Get the managing BatchOrchestrator instance.

Usage

BatchReactiveVal$get_orchestrator()

Returns

(R6 BatchOrchestrator) The BatchOrchestrator object.

Method clone()

The objects of this class are cloneable with this method.

Usage

BatchReactiveVal$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.