Create a batch reactive value

Creates a "smart" reactive value that coordinates with other reactive values in a batch to ensure atomic, validated updates. This is useful for managing interdependent inputs, like a set of cascading filters, where updating one value should immediately cause others to be re-validated.

Usage

batch_reactive_val(
  name,
  value = NULL,
  orchestrator = batch_orchestrator(),
  validation_fun = NULL
)

Arguments

name

(length-1 character) A unique character string to identify this reactive value within its batch.

value

(various) The initial value.

orchestrator

(BatchOrchestrator or batch_reactive_val) A BatchOrchestrator object created by batch_orchestrator(), or another batch_reactive_val that is already part of the desired batch. Defaults to creating a new, empty orchestrator.

validation_fun

(function or NULL) An optional function that defines the validation logic for this value. It must accept two arguments: its own current or proposed value, and a named list of all proposed values in the batch. The function must return the final, validated value based on the other values (usually either the input value or a default value).

Value

A function that behaves like a reactiveVal but with transactional guarantees, with class batch_reactive_val.