Create the User-Facing Getter/Setter Function

Create the user-facing function. Without an argument, the resulting function triggers the validation reactive. With an argument, the resulting function imperatively updates the internal state.

Usage

.create_vrv_function(state_rv, validation_rctv)

Arguments

state_rv

(reactiveVal) The shiny::reactiveVal() that holds the state.

validation_rctv

(reactive) The shiny::reactive() that performs the validation.

Value

A function that acts as a getter and setter.