Wrap an expression with error handling

A wrapper around rlang::try_fetch() that allows for cleanup code to be run before re-throwing an error with .shinybatch_abort().

Usage

.with_error_handling(
  expr,
  message,
  subclass,
  before_error = NULL,
  call = rlang::caller_env(),
  message_env = call,
  pass_parent = TRUE
)

Arguments

expr

(expression) The expression to evaluate.

message

(character) The message for the new error. Messages will be formatted with cli::cli_bullets().

subclass

(character) Class(es) to assign to the error. Will be prefixed by "shinybatch-error-".

before_error

(expression or NULL) An expression to run before aborting, typically for cleanup.

call

(environment) The execution environment to mention as the source of error messages.

message_env

(environment) The execution environment to use to evaluate variables in error messages.

pass_parent

(length-1 logical) Whether the triggering condition should be included in the error message.

Value

The result of expr if successful.