Create a validated reactive factor

A wrapper around validated_reactive_val() to help manage factors, especially in a cascading / dependent manner. This is a convenience function that constructs the validation_expr for you.

Usage

vrv_factor(
  levels,
  default,
  value = NULL,
  label = NULL,
  env = rlang::caller_env()
)

Arguments

levels

(expression) A reactive expression that returns a character vector of valid levels.

default

(expression) A reactive expression that returns a length-1 character vector to use as the default value when the current value is not in levels.

value

(various) The initial value.

label

(length-1 character or NULL) An optional label for the validated_reactive_value, used for debugging.

env

(environment) The environment in which to evaluate the validation_expr.

Value

A validated_reactive_val object.